requirements.txt
Flask==2.1.2
matplotlib==3.5.2
numpy==1.22.4
opencv_python==4.6.0.66
python-dotenv==0.20.0
Werkzeug==2.1.2
